+++ getwd.c
@@ -41,3 +41,11 @@ exit:
 
 	return ret;
 }
+
+char *__getwd_chk(char *buf, size_t buflen)
+{
+	__debug("%s(buf: %p, buflen: %zu)\n", __func__, buf, buflen);
+
+	/* Forward to another function */
+	return getwd(buf);
+}