+++ readlinkat.c
@@ -92,3 +92,13 @@ exit:
 
 	return ret;
 }
+
+ssize_t __readlinkat_chk(int fd, const char *path, char *buf, size_t pathlen,
+			 size_t bufsize)
+{
+	__debug("%s(path: '%s', buf: %p, pathlen: %zu, bufsize: %zu)\n",
+		__func__, path, buf, pathlen, bufsize);
+
+	/* Forward to another function */
+	return readlinkat(fd, path, buf, bufsize);
+}